Ignore SNAPSHOT versions in version ranges
Ignore SNAPSHOT releases when a dependency declares a version
range. Artifacts never expect a snapshot to show up when declaring
a version range and generally speaking it is not a problem, except
when using a mirror that mirrors both the release and snapshot versions
of an artifact. In that case, the metadata for the versions within a
range will contain both snapshot and release versions.

This change unconditionally filters out all SNAPSHOT versions from
version ranges.